+++ truco-front/src/App.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 
 function App() {
 
@@ -5,5 +6,7 @@ function App() {
     <h1 className="text-yellow-500">Hi</h1>
   )
 }
+
+
 
 export default App
